Close #352 - Change Effectful to fx and move to core / also move possible error syntax to core
