+++ cats-effect/src/main/scala-3/effectie/cats/syntax/all.scala
@@ -1,0 +1,6 @@
+package effectie.cats.syntax
+
+/** @author Kevin Lee
+  * @since 2021-10-23
+  */
+object all extends error
