+++ cats-effect/src/test/scala-2/effectie/cats/CanCatchSpec.scala
@@ -6,7 +6,7 @@ import cats.effect._
 import cats.instances.all._
 import cats.syntax.all._
 import effectie.SomeControlThrowable
-import effectie.cats.Effectful._
+import effectie.syntax.fx._
 import effectie.cats.Fx._
 import effectie.testing.types._
 import extras.concurrent.testing.ConcurrentSupport
