+++ cats-effect/src/test/scala-2/effectie/cats/CanRecoverSpec.scala
@@ -5,7 +5,7 @@ import cats.data.EitherT
 import cats.effect.IO
 import cats.instances.all._
 import cats.syntax.all._
-import effectie.cats.Effectful._
+import effectie.syntax.fx._
 import effectie.cats.FxCtor._
 import effectie.testing.types.SomeError
 import effectie.SomeControlThrowable
