+++ cats-effect/src/test/scala-2/effectie/cats/syntax/errorSpec.scala
@@ -4,9 +4,10 @@ import cats._
 import cats.data.EitherT
 import cats.effect._
 import cats.syntax.all._
-import effectie.cats.Effectful._
 import effectie.cats.Fx._
 import effectie.cats.syntax.error._
+import effectie.syntax.error._
+import effectie.syntax.fx._
 import effectie.testing.types._
 import effectie.{FxCtor, SomeControlThrowable}
 import extras.concurrent.testing.ConcurrentSupport
