+++ cats-effect/src/test/scala-3/effectie/syntax/fxSpec.scala
@@ -1,12 +1,11 @@
-package effectie.cats
+package effectie.syntax
 
 import cats.Id
 import cats.effect.IO
-import effectie.cats.CatsEffectRunner.TestContext
-import cats.effect.unsafe.IORuntime
-import effectie.cats.compat.CatsEffectIoCompatForFuture
+import effectie.cats.Fx.given
+import effectie.syntax.fx.*
 import effectie.testing.tools.*
-import effectie.testing.types.*
+import effectie.testing.types.SomeThrowableError
 import effectie.{Fx, FxCtor}
 import extras.concurrent.testing.ConcurrentSupport
 import extras.concurrent.testing.types.{ErrorLogger, WaitFor}
@@ -16,36 +15,33 @@ import hedgehog.runner.*
 /** @author Kevin Lee
   * @since 2021-05-16
   */
-object EffectfulSpec extends Properties {
+object fxSpec extends Properties {
   override def tests: List[Test] = List(
-    property("test Effectful.{effectOf, pureOf, unitOf} for IO", IoSpec.testAll),
-    property("test Effectful.effectOf[IO]", IoSpec.testEffectOf),
-    property("test Effectful.pureOf[IO]", IoSpec.testPureOf),
-    example("test Effectful.unitOf[IO]", IoSpec.testUnitOf),
-    example("test Effectful.errorOf[IO]", IoSpec.testErrorOf),
-    property("test Effectful.{effectOf, pureOf, unitOf} for Future", FutureSpec.testAll),
-    property("test Effectful.effectOf[Future]", FutureSpec.testEffectOf),
-    property("test Effectful.pureOf[Future]", FutureSpec.testPureOf),
-    example("test Effectful.unitOf[Future]", FutureSpec.testUnitOf),
-    example("test Effectful.errorOf[Future]", FutureSpec.testErrorOf),
-    property("test Effectful.{effectOf, pureOf, unitOf} for Id", IdSpec.testAll),
-    property("test Effectful.effectOf[Id]", IdSpec.testEffectOf),
-    property("test Effectful.pureOf[Id]", IdSpec.testPureOf),
-    example("test Effectful.unitOf[Id]", IdSpec.testUnitOf),
-    example("test Effectful.errorOf[Id]", IdSpec.testErrorOf)
+    property("test fx.{effectOf, pureOf, unitOf} for IO", IoSpec.testAll),
+    property("test fx.effectOf[IO]", IoSpec.testEffectOf),
+    property("test fx.pureOf[IO]", IoSpec.testPureOf),
+    example("test fx.unitOf[IO]", IoSpec.testUnitOf),
+    example("test fx.errorOf[IO]", IoSpec.testErrorOf),
+    property("test fx.{effectOf, pureOf, unitOf} for Future", FutureSpec.testAll),
+    property("test fx.effectOf[Future]", FutureSpec.testEffectOf),
+    property("test fx.pureOf[Future]", FutureSpec.testPureOf),
+    example("test fx.unitOf[Future]", FutureSpec.testUnitOf),
+    example("test fx.errorOf[Future]", FutureSpec.testErrorOf),
+    property("test fx.{effectOf, pureOf, unitOf} for Id", IdSpec.testAll),
+    property("test fx.effectOf[Id]", IdSpec.testEffectOf),
+    property("test fx.pureOf[Id]", IdSpec.testPureOf),
+    example("test fx.unitOf[Id]", IdSpec.testUnitOf),
+    example("test fx.errorOf[Id]", IdSpec.testErrorOf)
   )
-
-  import Effectful.*
 
   trait FxCtorClient[F[*]] {
     def eftOf[A](a: A): F[A]
     def of[A](a: A): F[A]
     def unit: F[Unit]
   }
-  object FxCtorClient      {
-    def apply[F[*]: FxCtorClient]: FxCtorClient[F]         = implicitly[FxCtorClient[F]]
-    implicit def eftClientF[F[*]: FxCtor]: FxCtorClient[F] = new FxCtorClientF[F]
-    final class FxCtorClientF[F[*]: FxCtor] extends FxCtorClient[F] {
+  object FxCtorClient {
+    def apply[F[*]: FxCtorClient]: FxCtorClient[F] = summon[FxCtorClient[F]]
+    given eftClientF[F[*]: FxCtor]: FxCtorClient[F] with {
       override def eftOf[A](a: A): F[A] = effectOf(a)
       override def of[A](a: A): F[A]    = pureOf(a)
       override def unit: F[Unit]        = unitOf
@@ -57,11 +53,10 @@ object EffectfulSpec extends Properties {
     def of[A](a: A): F[A]
     def unit: F[Unit]
   }
-  object FxClient      {
-    def apply[F[*]: FxClient]: FxClient[F]         =
-      implicitly[FxClient[F]]
-    implicit def eftClientF[F[*]: Fx]: FxClient[F] = new FxClientF[F]
-    final class FxClientF[F[*]: Fx] extends FxClient[F] {
+  object FxClient {
+    def apply[F[*]: FxClient]: FxClient[F] =
+      summon[FxClient[F]]
+    given eftClientF[F[*]: Fx]: FxClient[F] with {
       override def eftOf[A](a: A): F[A] = effectOf(a)
       override def of[A](a: A): F[A]    = pureOf(a)
       override def unit: F[Unit]        = unitOf
@@ -69,10 +64,6 @@ object EffectfulSpec extends Properties {
   }
 
   object IoSpec {
-    import effectie.cats.Fx.given
-
-    val compat          = new CatsEffectIoCompatForFuture
-    given rt: IORuntime = testing.IoAppUtils.runtime(compat.es)
 
     def testAll: Property = for {
       before <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("before")
@@ -80,8 +71,8 @@ object EffectfulSpec extends Properties {
     } yield {
       var actual                  = before
       var actual2                 = before
-      val testBefore              = (actual ==== before).log(s"actual testBefore should be $before but was $actual")
-      val testBefore2             = (actual2 ==== before).log(s"actual2 testBefore2 should be $before but was $actual2")
+      val testBefore              = actual ==== before
+      val testBefore2             = actual2 ==== before
       val eftClient               = FxCtorClient[IO]
       val effectConstructorClient = FxClient[IO]
       val io                      =
@@ -95,23 +86,18 @@ object EffectfulSpec extends Properties {
           _  <- eftClient.unit
           _  <- effectConstructorClient.unit
         } yield ()
-      val testBeforeRun           = (actual ==== before).log(s"actual testBeforeRun should be $before but was $actual")
-      val testBeforeRun2          = (actual2 ==== before).log(s"actual2 testBeforeRun2 should be $before but was $actual2")
-
-      import CatsEffectRunner.*
-      given ticket: Ticker = Ticker(TestContext())
-
-      val runResult     = io.completeAs(())
-      val testAfterRun  = (actual ==== after).log(s"actual testAfterRun should be $after but was $actual")
-      val testAfterRun2 = (actual2 ==== after).log(s"actual2 testAfterRun2 should be $after but was $actual2")
+      val testBeforeRun           = actual ==== before
+      val testBeforeRun2          = actual2 ==== before
+      io.unsafeRunSync()
+      val testAfterRun            = actual ==== after
+      val testAfterRun2           = actual2 ==== after
       Result.all(
         List(
           testBefore.log("testBefore"),
-          testBefore2.log("testBefore2"),
           testBeforeRun.log("testBeforeRun"),
-          testBeforeRun2.log("testBeforeRun2"),
-          runResult,
           testAfterRun.log("testAfterRun"),
+          testBefore2.log("testBefore2"),
+          testBeforeRun2.log("testBeforeRun2"),
           testAfterRun2.log("testAfterRun2")
         )
       )
@@ -122,20 +108,15 @@ object EffectfulSpec extends Properties {
       after  <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).map(_ + before).log("after")
     } yield {
       var actual        = before
-      val testBefore    = (actual ==== before).log(s"actual before should be $before but was $actual")
+      val testBefore    = actual ==== before
       val io            = effectOf[IO]({ actual = after; () })
-      val testBeforeRun = (actual ==== before).log(s"actual beforeRun should be $before but was $actual")
-
-      import CatsEffectRunner.*
-      given ticket: Ticker = Ticker(TestContext())
-
-      val runResult    = io.completeAs(())
-      val testAfterRun = (actual ==== after).log(s"actual afterRun should be $after but was $actual")
+      val testBeforeRun = actual ==== before
+      io.unsafeRunSync()
+      val testAfterRun  = actual ==== after
       Result.all(
         List(
           testBefore.log("testBefore"),
           testBeforeRun.log("testBeforeRun"),
-          runResult,
           testAfterRun.log("testAfterRun")
         )
       )
@@ -145,45 +126,38 @@ object EffectfulSpec extends Properties {
       before <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("before")
       after  <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).map(_ + before).log("after")
     } yield {
-      import CatsEffectRunner.*
-      given ticket: Ticker = Ticker(TestContext())
-
       var actual        = before
-      val testBefore    = (actual ==== before).log(s"actual before should be $before but was $actual")
+      val testBefore    = actual ==== before
       val io            = pureOf[IO]({ actual = after; () })
-      val testBeforeRun = (actual ==== after).log(s"actual beforeRun should be $after but was $actual")
-
-      val runResult = io.completeAs(())
-
-      val testAfterRun = (actual ==== after).log(s"actual afterRun should be $after but was $actual")
+      val testBeforeRun = actual ==== after
+      io.unsafeRunSync()
+      val testAfterRun  = actual ==== after
       Result.all(
         List(
           testBefore.log("testBefore"),
           testBeforeRun.log("testBeforeRun"),
-          runResult,
           testAfterRun.log("testAfterRun")
         )
       )
     }
 
     def testUnitOf: Result = {
-      import CatsEffectRunner.*
-      given ticket: Ticker = Ticker(TestContext())
-
       val io             = unitOf[IO]
       val expected: Unit = ()
-      io.completeAs(expected)
+      val actual: Unit   = io.unsafeRunSync()
+      actual ==== expected
     }
 
     def testErrorOf: Result = {
       val expectedMessage = "This is a throwable test error."
       val expectedError   = SomeThrowableError.message(expectedMessage)
 
-      import CatsEffectRunner.*
-      given ticket: Ticker = Ticker(TestContext())
-
       val io = errorOf[IO][Unit](expectedError)
-      io.expectError(expectedError)
+
+      expectThrowable(
+        io.unsafeRunSync(),
+        expectedError
+      )
     }
 
   }
@@ -194,7 +168,7 @@ object EffectfulSpec extends Properties {
     import scala.concurrent.duration.*
     import scala.concurrent.{ExecutionContext, Future}
 
-    private given errorLogger: ErrorLogger[Throwable] = ErrorLogger.printlnDefaultErrorLogger
+    private implicit val errorLogger: ErrorLogger[Throwable] = ErrorLogger.printlnDefaultErrorLogger
 
     private val waitFor = WaitFor(1.second)
 
@@ -203,7 +177,8 @@ object EffectfulSpec extends Properties {
       after  <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).map(_ + before).log("after")
     } yield {
       given executorService: ExecutorService = Executors.newFixedThreadPool(1)
-      given ExecutionContext                 = ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
+      given ec: ExecutionContext             =
+        ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
 
       var actual                  = before
       var actual2                 = before
@@ -240,7 +215,8 @@ object EffectfulSpec extends Properties {
       after  <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).map(_ + before).log("after")
     } yield {
       given executorService: ExecutorService = Executors.newFixedThreadPool(1)
-      given ExecutionContext                 = ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
+      given ec: ExecutionContext             =
+        ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
 
       var actual               = before
       val testBefore           = actual ==== before
@@ -260,7 +236,8 @@ object EffectfulSpec extends Properties {
       after  <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).map(_ + before).log("after")
     } yield {
       given executorService: ExecutorService = Executors.newFixedThreadPool(1)
-      given ExecutionContext                 = ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
+      given ec: ExecutionContext             =
+        ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
 
       var actual       = before
       val testBefore   = actual ==== before
@@ -277,10 +254,11 @@ object EffectfulSpec extends Properties {
 
     def testUnitOf: Result = {
       given executorService: ExecutorService = Executors.newFixedThreadPool(1)
-      given ExecutionContext                 = ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
+      given ec: ExecutionContext             =
+        ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
       val future                             = unitOf[Future]
       val expected: Unit                     = ()
-      val actual: Unit                       = ConcurrentSupport.futureToValueAndTerminate(executorService, waitFor)(future)
+      val actual: Unit = ConcurrentSupport.futureToValueAndTerminate(executorService, waitFor)(future)
       actual ==== expected
     }
 
@@ -289,16 +267,19 @@ object EffectfulSpec extends Properties {
       val expectedError   = SomeThrowableError.message(expectedMessage)
 
       given executorService: ExecutorService = Executors.newFixedThreadPool(1)
-      given ExecutionContext                 = ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
+      given ec: ExecutionContext             =
+        ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
+      val future                             = errorOf[Future][Unit](expectedError)
 
-      val future = errorOf[Future][Unit](expectedError)
-      expectThrowable(ConcurrentSupport.futureToValueAndTerminate(executorService, waitFor)(future), expectedError)
+      expectThrowable(
+        ConcurrentSupport.futureToValueAndTerminate(executorService, waitFor)(future),
+        expectedError
+      )
     }
 
   }
 
   object IdSpec {
-    import effectie.cats.Fx.given
 
     def testAll: Property = for {
       before <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("before")
