+++ cats-effect3/src/main/scala-2/effectie/cats/syntax/all.scala
@@ -1,0 +1,7 @@
+package effectie.cats.syntax
+
+/**
+ * @author Kevin Lee
+ * @since 2022-01-19
+ */
+object all extends error
