+++ cats-effect3/src/test/scala-2/effectie/cats/CanHandleErrorSpec.scala
@@ -7,7 +7,7 @@ import cats.effect.unsafe.IORuntime
 import cats.instances.all._
 import cats.syntax.all._
 import effectie.cats.CanHandleError._
-import effectie.cats.Effectful._
+import effectie.syntax.fx._
 import effectie.cats.FxCtor._
 import effectie.testing.types.SomeError
 import effectie.SomeControlThrowable
