+++ cats-effect3/src/test/scala-2/effectie/cats/CanRecoverSpec.scala
@@ -8,7 +8,7 @@ import cats.instances.all._
 import cats.syntax.all._
 import effectie.cats.CanRecover._
 import effectie.cats.FxCtor._
-import effectie.cats.Effectful._
+import effectie.syntax.fx._
 import effectie.cats.compat.CatsEffectIoCompatForFuture
 import effectie.testing.types.SomeError
 import effectie.SomeControlThrowable
