+++ cats-effect3/src/test/scala-2/effectie/syntax/fxSpec.scala
@@ -1,47 +1,44 @@
-package effectie.monix
+package effectie.syntax
 
 import cats.Id
-import effectie.testing.tools._
+import cats.effect.IO
+import cats.effect.unsafe.IORuntime
+import effectie.cats.Fx._
+import effectie.cats.compat.CatsEffectIoCompatForFuture
+import effectie.cats.{CatsEffectRunner, testing}
+import effectie.syntax.fx._
+import effectie.testing.tools.{dropResult, expectThrowable}
 import effectie.testing.types.SomeThrowableError
+import effectie.{Fx, FxCtor}
 import extras.concurrent.testing.ConcurrentSupport
 import extras.concurrent.testing.types.{ErrorLogger, WaitFor}
 import hedgehog._
 import hedgehog.runner._
-import monix.eval.Task
 
 /** @author Kevin Lee
   * @since 2021-05-16
   */
-object EffectfulSpec extends Properties {
+object fxSpec extends Properties {
+
   private implicit val errorLogger: ErrorLogger[Throwable] = ErrorLogger.printlnDefaultErrorLogger
 
-  type Fx[F[_]]     = effectie.Fx[F]
-  type FxCtor[F[_]] = effectie.FxCtor[F]
-
   override def tests: List[Test] = List(
-    property("test Effectful.{effectOf, pureOf, unitOf} for Task", TaskSpec.testAll),
-    property("test Effectful.effectOf[Task]", TaskSpec.testEffectOf),
-    property("test Effectful.pureOf[Task]", TaskSpec.testPureOf),
-    example("test Effectful.unitOf[Task]", TaskSpec.testUnitOf),
-    example("test Effectful.errorOf[Task]", TaskSpec.testErrorOf),
-    property("test Effectful.{effectOf, pureOf, unitOf} for IO", TaskSpec.testAll),
-    property("test Effectful.effectOf[IO]", IoSpec.testEffectOf),
-    property("test Effectful.pureOf[IO]", IoSpec.testPureOf),
-    example("test Effectful.unitOf[IO]", IoSpec.testUnitOf),
-    example("test Effectful.errorOf[IO]", IoSpec.testErrorOf),
-    property("test Effectful.{effectOf, pureOf, unitOf} for Future", FutureSpec.testAll),
-    property("test Effectful.effectOf[Future]", FutureSpec.testEffectOf),
-    property("test Effectful.pureOf[Future]", FutureSpec.testPureOf),
-    example("test Effectful.unitOf[Future]", FutureSpec.testUnitOf),
-    example("test Effectful.errorOf[Future]", FutureSpec.testErrorOf),
-    property("test Effectful.{effectOf, pureOf, unitOf} for Id", IdSpec.testAll),
-    property("test Effectful.effectOf[Id]", IdSpec.testEffectOf),
-    property("test Effectful.pureOf[Id]", IdSpec.testPureOf),
-    example("test Effectful.unitOf[Id]", IdSpec.testUnitOf),
-    example("test Effectful.errorOf[Id]", IdSpec.testErrorOf)
+    property("test fx.{effectOf, pureOf, unitOf} for IO", IoSpec.testAll),
+    property("test fx.effectOf[IO]", IoSpec.testEffectOf),
+    property("test fx.pureOf[IO]", IoSpec.testPureOf),
+    example("test fx.unitOf[IO]", IoSpec.testUnitOf),
+    example("test fx.errorOf[IO]", IoSpec.testErrorOf),
+    property("test fx.{effectOf, pureOf, unitOf} for Future", FutureSpec.testAll),
+    property("test fx.effectOf[Future]", FutureSpec.testEffectOf),
+    property("test fx.pureOf[Future]", FutureSpec.testPureOf),
+    example("test fx.unitOf[Future]", FutureSpec.testUnitOf),
+    example("test fx.errorOf[Future]", FutureSpec.testErrorOf),
+    property("test fx.{effectOf, pureOf, unitOf} for Id", IdSpec.testAll),
+    property("test fx.effectOf[Id]", IdSpec.testEffectOf),
+    property("test fx.pureOf[Id]", IdSpec.testPureOf),
+    example("test fx.unitOf[Id]", IdSpec.testUnitOf),
+    example("test fx.errorOf[Id]", IdSpec.testErrorOf)
   )
-
-  import Effectful._
 
   trait FxCtorClient[F[_]] {
     def eftOf[A](a: A): F[A]
@@ -53,10 +50,10 @@ object EffectfulSpec extends Properties {
     def apply[F[_]: FxCtorClient]: FxCtorClient[F]         = implicitly[FxCtorClient[F]]
     implicit def eftClientF[F[_]: FxCtor]: FxCtorClient[F] = new FxCtorClientF[F]
     final class FxCtorClientF[F[_]: FxCtor] extends FxCtorClient[F] {
-      override def eftOf[A](a: A): F[A]                 = effectOf(a)
-      override def of[A](a: A): F[A]                    = pureOf(a)
-      override def unit: F[Unit]                        = unitOf
-      override def errOf[A](throwable: Throwable): F[A] = errorOf(throwable)
+      override def eftOf[A](a: A): F[A]                 = effectOf[F](a)
+      override def of[A](a: A): F[A]                    = pureOf[F](a)
+      override def unit: F[Unit]                        = unitOf[F]
+      override def errOf[A](throwable: Throwable): F[A] = errorOf[F](throwable)
     }
   }
 
@@ -71,122 +68,17 @@ object EffectfulSpec extends Properties {
       implicitly[FxClient[F]]
     implicit def eftClientF[F[_]: Fx]: FxClient[F] = new FxClientF[F]
     final class FxClientF[F[_]: Fx] extends FxClient[F] {
-      override def eftOf[A](a: A): F[A]                 = effectOf(a)
-      override def of[A](a: A): F[A]                    = pureOf(a)
-      override def unit: F[Unit]                        = unitOf
-      override def errOf[A](throwable: Throwable): F[A] = errorOf(throwable)
+      override def eftOf[A](a: A): F[A]                 = effectOf[F](a)
+      override def of[A](a: A): F[A]                    = pureOf[F](a)
+      override def unit: F[Unit]                        = unitOf[F]
+      override def errOf[A](throwable: Throwable): F[A] = errorOf[F](throwable)
     }
-  }
-
-  object TaskSpec {
-    import effectie.monix.Fx._
-
-    import monix.execution.Scheduler.Implicits.global
-
-    @SuppressWarnings(Array("org.wartremover.warts.Any", "org.wartremover.warts.Nothing"))
-    def testAll: Property = for {
-      before <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("before")
-      after  <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).map(_ + before).log("after")
-    } yield {
-      @SuppressWarnings(Array("org.wartremover.warts.Var"))
-      var actual                  = before
-      @SuppressWarnings(Array("org.wartremover.warts.Var"))
-      var actual2                 = before
-      val testBefore              = actual ==== before
-      val testBefore2             = actual2 ==== before
-      val eftClient               = FxCtorClient[Task]
-      val effectConstructorClient = FxClient[Task]
-      val io                      =
-        for {
-          _  <- effectOf[Task]({ actual = after; () })
-          _  <- pureOf[Task]({ actual2 = after; () })
-          n  <- eftClient.eftOf(1)
-          n2 <- eftClient.of(n)
-          i  <- effectConstructorClient.eftOf(1)
-          i2 <- effectConstructorClient.of(1)
-          _  <- eftClient.unit
-          _  <- effectConstructorClient.unit
-        } yield ()
-      val testBeforeRun           = actual ==== before
-      val testBeforeRun2          = actual2 ==== before
-      io.runSyncUnsafe()
-      val testAfterRun            = actual ==== after
-      val testAfterRun2           = actual2 ==== after
-      Result.all(
-        List(
-          testBefore.log("testBefore"),
-          testBeforeRun.log("testBeforeRun"),
-          testAfterRun.log("testAfterRun"),
-          testBefore2.log("testBefore2"),
-          testBeforeRun2.log("testBeforeRun2"),
-          testAfterRun2.log("testAfterRun2")
-        )
-      )
-    }
-
-    @SuppressWarnings(Array("org.wartremover.warts.Any", "org.wartremover.warts.Nothing"))
-    def testEffectOf: Property = for {
-      before <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("before")
-      after  <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).map(_ + before).log("after")
-    } yield {
-      @SuppressWarnings(Array("org.wartremover.warts.Var"))
-      var actual        = before
-      val testBefore    = actual ==== before
-      val io            = effectOf[Task]({ actual = after; () })
-      val testBeforeRun = actual ==== before
-      io.runSyncUnsafe()
-      val testAfterRun  = actual ==== after
-      Result.all(
-        List(
-          testBefore.log("testBefore"),
-          testBeforeRun.log("testBeforeRun"),
-          testAfterRun.log("testAfterRun")
-        )
-      )
-    }
-
-    def testPureOf: Property = for {
-      before <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("before")
-      after  <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).map(_ + before).log("after")
-    } yield {
-      @SuppressWarnings(Array("org.wartremover.warts.Var"))
-      var actual        = before
-      val testBefore    = actual ==== before
-      val io            = pureOf[Task]({ actual = after; () })
-      val testBeforeRun = actual ==== after
-      io.runSyncUnsafe()
-      val testAfterRun  = actual ==== after
-      Result.all(
-        List(
-          testBefore.log("testBefore"),
-          testBeforeRun.log("testBeforeRun"),
-          testAfterRun.log("testAfterRun")
-        )
-      )
-    }
-
-    def testUnitOf: Result = {
-      val io             = unitOf[Task]
-      val expected: Unit = ()
-      val actual: Unit   = io.runSyncUnsafe()
-      actual ==== expected
-    }
-
-    @SuppressWarnings(Array("org.wartremover.warts.Any", "org.wartremover.warts.Nothing"))
-    def testErrorOf: Result = {
-      val expectedMessage = "This is a throwable test error."
-      val expectedError   = SomeThrowableError.message(expectedMessage)
-
-      val io = errorOf[Task][Unit](expectedError)
-      expectThrowable(io.runSyncUnsafe(), expectedError)
-    }
-
   }
 
   object IoSpec {
-    import effectie.monix.Fx._
 
-    import cats.effect.IO
+    val compat                 = new CatsEffectIoCompatForFuture
+    implicit val rt: IORuntime = testing.IoAppUtils.runtime(compat.es)
 
     @SuppressWarnings(Array("org.wartremover.warts.Any", "org.wartremover.warts.Nothing"))
     def testAll: Property = for {
@@ -214,7 +106,9 @@ object EffectfulSpec extends Properties {
         } yield ()
       val testBeforeRun           = actual ==== before
       val testBeforeRun2          = actual2 ==== before
-      io.unsafeRunSync()
+      import CatsEffectRunner._
+      implicit val ticket: Ticker = Ticker(TestContext())
+      val runResult               = io.completeAs(())
       val testAfterRun            = actual ==== after
       val testAfterRun2           = actual2 ==== after
       Result.all(
@@ -223,6 +117,7 @@ object EffectfulSpec extends Properties {
           testBefore2.log("testBefore2"),
           testBeforeRun.log("testBeforeRun"),
           testBeforeRun2.log("testBeforeRun2"),
+          runResult,
           testAfterRun.log("testAfterRun"),
           testAfterRun2.log("testAfterRun2")
         )
@@ -239,12 +134,17 @@ object EffectfulSpec extends Properties {
       val testBefore    = actual ==== before
       val io            = effectOf[IO]({ actual = after; () })
       val testBeforeRun = actual ==== before
-      io.unsafeRunSync()
-      val testAfterRun  = actual ==== after
+
+      import CatsEffectRunner._
+      implicit val ticket: Ticker = Ticker(TestContext())
+      val runResult               = io.completeAs(())
+
+      val testAfterRun = actual ==== after
       Result.all(
         List(
           testBefore.log("testBefore"),
           testBeforeRun.log("testBeforeRun"),
+          runResult,
           testAfterRun.log("testAfterRun")
         )
       )
@@ -259,12 +159,17 @@ object EffectfulSpec extends Properties {
       val testBefore    = actual ==== before
       val io            = pureOf[IO]({ actual = after; () })
       val testBeforeRun = actual ==== after
-      io.unsafeRunSync()
-      val testAfterRun  = actual ==== after
+
+      import CatsEffectRunner._
+      implicit val ticket: Ticker = Ticker(TestContext())
+      val runResult               = io.completeAs(())
+
+      val testAfterRun = actual ==== after
       Result.all(
         List(
           testBefore.log("testBefore"),
           testBeforeRun.log("testBeforeRun"),
+          runResult,
           testAfterRun.log("testAfterRun")
         )
       )
@@ -273,8 +178,11 @@ object EffectfulSpec extends Properties {
     def testUnitOf: Result = {
       val io             = unitOf[IO]
       val expected: Unit = ()
-      val actual: Unit   = io.unsafeRunSync()
-      actual ==== expected
+
+      import CatsEffectRunner._
+      implicit val ticket: Ticker = Ticker(TestContext())
+
+      io.completeAs(expected)
     }
 
     @SuppressWarnings(Array("org.wartremover.warts.Any", "org.wartremover.warts.Nothing"))
@@ -283,7 +191,11 @@ object EffectfulSpec extends Properties {
       val expectedError   = SomeThrowableError.message(expectedMessage)
 
       val io = errorOf[IO][Unit](expectedError)
-      expectThrowable(io.unsafeRunSync(), expectedError)
+
+      import CatsEffectRunner._
+      implicit val ticket: Ticker = Ticker(TestContext())
+
+      io.expectError(expectedError)
     }
 
   }
@@ -301,7 +213,8 @@ object EffectfulSpec extends Properties {
       after  <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).map(_ + before).log("after")
     } yield {
       implicit val executorService: ExecutorService = Executors.newFixedThreadPool(1)
-      implicit val ec: ExecutionContext             = ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
+      implicit val ec: ExecutionContext             =
+        ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
 
       @SuppressWarnings(Array("org.wartremover.warts.Var"))
       var actual                  = before
@@ -340,7 +253,8 @@ object EffectfulSpec extends Properties {
       after  <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).map(_ + before).log("after")
     } yield {
       implicit val executorService: ExecutorService = Executors.newFixedThreadPool(1)
-      implicit val ec: ExecutionContext             = ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
+      implicit val ec: ExecutionContext             =
+        ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
 
       @SuppressWarnings(Array("org.wartremover.warts.Var"))
       var actual               = before
@@ -361,7 +275,8 @@ object EffectfulSpec extends Properties {
       after  <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).map(_ + before).log("after")
     } yield {
       implicit val executorService: ExecutorService = Executors.newFixedThreadPool(1)
-      implicit val ec: ExecutionContext             = ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
+      implicit val ec: ExecutionContext             =
+        ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
 
       @SuppressWarnings(Array("org.wartremover.warts.Var"))
       var actual       = before
@@ -379,10 +294,11 @@ object EffectfulSpec extends Properties {
 
     def testUnitOf: Result = {
       implicit val executorService: ExecutorService = Executors.newFixedThreadPool(1)
-      implicit val ec: ExecutionContext             = ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
+      implicit val ec: ExecutionContext             =
+        ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
       val future                                    = unitOf[Future]
       val expected: Unit                            = ()
-      val actual: Unit                              = ConcurrentSupport.futureToValueAndTerminate(executorService, waitFor)(future)
+      val actual: Unit = ConcurrentSupport.futureToValueAndTerminate(executorService, waitFor)(future)
       actual ==== expected
     }
 
@@ -391,7 +307,8 @@ object EffectfulSpec extends Properties {
       val expectedError   = SomeThrowableError.message(expectedMessage)
 
       implicit val executorService: ExecutorService = Executors.newFixedThreadPool(1)
-      implicit val ec: ExecutionContext             = ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
+      implicit val ec: ExecutionContext             =
+        ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
 
       val future = errorOf[Future][Unit](expectedError)
       expectThrowable(ConcurrentSupport.futureToValueAndTerminate(executorService, waitFor)(future), expectedError)
@@ -400,7 +317,6 @@ object EffectfulSpec extends Properties {
   }
 
   object IdSpec {
-    import effectie.monix.Fx._
 
     def testAll: Property = for {
       before <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("before")
