+++ cats-effect3/src/test/scala-3/effectie/cats/CanRecoverSpec.scala
@@ -7,7 +7,7 @@ import effectie.cats.CatsEffectRunner.TestContext
 import cats.effect.unsafe.IORuntime
 import cats.instances.all.*
 import cats.syntax.all.*
-import effectie.cats.Effectful.*
+import effectie.syntax.fx.*
 import effectie.cats.compat.CatsEffectIoCompatForFuture
 import effectie.testing.types.SomeError
 import effectie.{FxCtor, SomeControlThrowable}
