+++ cats-effect3/src/test/scala-3/effectie/cats/syntax/errorSpec.scala
@@ -1,17 +1,18 @@
 package effectie.cats.syntax
 
-import cats.{Functor, Id}
 import cats.data.EitherT
-import cats.syntax.all.*
 import cats.effect.IO
-import effectie.cats.CatsEffectRunner.TestContext
 import cats.effect.unsafe.IORuntime
-import effectie.{Fx, SomeControlThrowable}
-import effectie.cats.{CanCatch, FxCtor, testing}
-import effectie.testing.types.*
-import effectie.cats.Effectful.*
+import cats.syntax.all.*
+import cats.{Functor, Id}
+import effectie.cats.CatsEffectRunner.TestContext
 import effectie.cats.compat.CatsEffectIoCompatForFuture
 import effectie.cats.syntax.error.*
+import effectie.cats.{CanCatch, FxCtor, testing}
+import effectie.syntax.error.*
+import effectie.syntax.fx.*
+import effectie.testing.types.*
+import effectie.{Fx, SomeControlThrowable}
 import extras.concurrent.testing.ConcurrentSupport
 import extras.concurrent.testing.types.{ErrorLogger, WaitFor}
 import hedgehog.*
@@ -1109,8 +1110,8 @@ object CanHandleErrorSyntaxSpec {
     effectOf[F](a)
 
   object IoSpec {
-    import effectie.cats.Fx.given
     import effectie.cats.CanHandleError.ioCanHandleError
+    import effectie.cats.Fx.given
 
     def testCanHandleError_IO_handleNonFatalWithShouldHandleNonFatalWith: Result = {
 
@@ -2024,8 +2025,8 @@ object CanHandleErrorSyntaxSpec {
   }
 
   object IdSpec {
-    import effectie.cats.Fx.given
     import effectie.cats.CanHandleError.idCanHandleError
+    import effectie.cats.Fx.given
 
     def testCanHandleError_Id_handleNonFatalWithShouldHandleNonFatalWith: Result = {
 
@@ -2769,8 +2770,8 @@ object CanRecoverSyntaxSpec {
     effectOf[F](a)
 
   object IOSpec {
-    import effectie.cats.Fx.given
     import effectie.cats.CanRecover.given
+    import effectie.cats.Fx.given
 
     def testCanRecover_IO_recoverFromNonFatalWithShouldRecoverFromNonFatal: Result = {
 
@@ -3780,8 +3781,8 @@ object CanRecoverSyntaxSpec {
   }
 
   object IdSpec {
-    import effectie.cats.Fx.given
     import effectie.cats.CanRecover.given
+    import effectie.cats.Fx.given
 
     def testCanRecover_Id_recoverFromNonFatalWithShouldRecoverFromNonFatal: Result = {
