+++ core/src/main/scala-2/effectie/syntax/all.scala
@@ -1,0 +1,7 @@
+package effectie.syntax
+
+/**
+ * @author Kevin Lee
+ * @since 2022-01-18
+ */
+object all extends fx with error
