+++ core/src/main/scala-3/effectie/syntax/all.scala
@@ -1,0 +1,8 @@
+package effectie.syntax
+
+/**
+ * @author Kevin Lee
+ * @since 2022-01-17
+ */
+object all extends fx with error
+
