+++ effectie-monix/src/main/scala/effectie/monix/syntax/all.scala
@@ -1,0 +1,7 @@
+package effectie.monix.syntax
+
+/**
+ * @author Kevin Lee
+ * @since 2022-01-19
+ */
+object all extends error
