+++ effectie-monix/src/test/scala/effectie/monix/CanHandleErrorSpec.scala
@@ -5,11 +5,11 @@ import cats.data.EitherT
 import cats.effect.IO
 import cats.instances.all._
 import cats.syntax.all._
-import effectie.monix.CanHandleError._
-import effectie.monix.Effectful._
-import effectie.monix.FxCtor._
-import effectie.testing.types.SomeError
 import effectie.SomeControlThrowable
+import effectie.monix.CanHandleError._
+import effectie.monix.FxCtor._
+import effectie.syntax.fx._
+import effectie.testing.types.SomeError
 import extras.concurrent.testing.ConcurrentSupport
 import extras.concurrent.testing.types.{ErrorLogger, WaitFor}
 import hedgehog._
