+++ effectie-monix/src/test/scala/effectie/monix/CatchingSpec.scala
@@ -6,7 +6,7 @@ import cats.data.EitherT
 import cats.effect.IO
 import cats.instances.all._
 import cats.syntax.all._
-import effectie.monix.Effectful._
+import effectie.syntax.fx._
 import effectie.testing.types.SomeError
 import effectie.SomeControlThrowable
 import extras.concurrent.testing.ConcurrentSupport
