+++ effectie-monix/src/test/scala/effectie/syntax/fxSpec.scala
@@ -1,56 +1,60 @@
-package effectie.cats
+package effectie.syntax
 
-import cats.Id
-import cats.effect.IO
-import effectie.testing.tools._
+import cats._
+import effectie.monix.Fx._
+import effectie.syntax.fx._
+import effectie.testing.tools.{dropResult, expectThrowable}
 import effectie.testing.types.SomeThrowableError
+import effectie.{Fx, FxCtor}
 import extras.concurrent.testing.ConcurrentSupport
 import extras.concurrent.testing.types.{ErrorLogger, WaitFor}
 import hedgehog._
 import hedgehog.runner._
+import monix.eval.Task
 
 /** @author Kevin Lee
   * @since 2021-05-16
   */
-object EffectfulSpec extends Properties {
-  implicit val errorLogger: ErrorLogger[Throwable] = ErrorLogger.printlnDefaultErrorLogger
-
-  type Fx[F[_]] = effectie.Fx[F]
-  val Fx: effectie.Fx.type = effectie.Fx
-  type FxCtor[F[_]] = effectie.FxCtor[F]
+object fxSpec extends Properties {
+  private implicit val errorLogger: ErrorLogger[Throwable] = ErrorLogger.printlnDefaultErrorLogger
 
   override def tests: List[Test] = List(
-    property("test Effectful.{effectOf, pureOf, unitOf} for IO", IoSpec.testAll),
-    property("test Effectful.effectOf[IO]", IoSpec.testEffectOf),
-    property("test Effectful.pureOf[IO]", IoSpec.testPureOf),
-    example("test Effectful.unitOf[IO]", IoSpec.testUnitOf),
-    example("test Effectful.errorOf[IO]", IoSpec.testErrorOf),
-    property("test Effectful.{effectOf, pureOf, unitOf} for Future", FutureSpec.testAll),
-    property("test Effectful.effectOf[Future]", FutureSpec.testEffectOf),
-    property("test Effectful.pureOf[Future]", FutureSpec.testPureOf),
-    example("test Effectful.unitOf[Future]", FutureSpec.testUnitOf),
-    example("test Effectful.errorOf[Future]", FutureSpec.testErrorOf),
-    property("test Effectful.{effectOf, pureOf, unitOf} for Id", IdSpec.testAll),
-    property("test Effectful.effectOf[Id]", IdSpec.testEffectOf),
-    property("test Effectful.pureOf[Id]", IdSpec.testPureOf),
-    example("test Effectful.unitOf[Id]", IdSpec.testUnitOf),
-    example("test Effectful.errorOf[Id]", IdSpec.testErrorOf)
+    property("test fx.{effectOf, pureOf, unitOf} for Task", TaskSpec.testAll),
+    property("test fx.effectOf[Task]", TaskSpec.testEffectOf),
+    property("test fx.pureOf[Task]", TaskSpec.testPureOf),
+    example("test fx.unitOf[Task]", TaskSpec.testUnitOf),
+    example("test fx.errorOf[Task]", TaskSpec.testErrorOf),
+    property("test fx.{effectOf, pureOf, unitOf} for IO", TaskSpec.testAll),
+    property("test fx.effectOf[IO]", IoSpec.testEffectOf),
+    property("test fx.pureOf[IO]", IoSpec.testPureOf),
+    example("test fx.unitOf[IO]", IoSpec.testUnitOf),
+    example("test fx.errorOf[IO]", IoSpec.testErrorOf),
+    property("test fx.{effectOf, pureOf, unitOf} for Future", FutureSpec.testAll),
+    property("test fx.effectOf[Future]", FutureSpec.testEffectOf),
+    property("test fx.pureOf[Future]", FutureSpec.testPureOf),
+    example("test fx.unitOf[Future]", FutureSpec.testUnitOf),
+    example("test fx.errorOf[Future]", FutureSpec.testErrorOf),
+    property("test fx.{effectOf, pureOf, unitOf} for Id", IdSpec.testAll),
+    property("test fx.effectOf[Id]", IdSpec.testEffectOf),
+    property("test fx.pureOf[Id]", IdSpec.testPureOf),
+    example("test fx.unitOf[Id]", IdSpec.testUnitOf),
+    example("test fx.errorOf[Id]", IdSpec.testErrorOf)
   )
-
-  import Effectful._
 
   trait FxCtorClient[F[_]] {
     def eftOf[A](a: A): F[A]
     def of[A](a: A): F[A]
     def unit: F[Unit]
+    def errOf[A](throwable: Throwable): F[A]
   }
   object FxCtorClient {
     def apply[F[_]: FxCtorClient]: FxCtorClient[F]         = implicitly[FxCtorClient[F]]
     implicit def eftClientF[F[_]: FxCtor]: FxCtorClient[F] = new FxCtorClientF[F]
     final class FxCtorClientF[F[_]: FxCtor] extends FxCtorClient[F] {
-      override def eftOf[A](a: A): F[A] = effectOf(a)
-      override def of[A](a: A): F[A]    = pureOf(a)
-      override def unit: F[Unit]        = unitOf
+      override def eftOf[A](a: A): F[A]                 = effectOf[F](a)
+      override def of[A](a: A): F[A]                    = pureOf[F](a)
+      override def unit: F[Unit]                        = unitOf[F]
+      override def errOf[A](throwable: Throwable): F[A] = errorOf[F](throwable)
     }
   }
 
@@ -58,20 +62,126 @@ object EffectfulSpec extends Properties {
     def eftOf[A](a: A): F[A]
     def of[A](a: A): F[A]
     def unit: F[Unit]
+    def errOf[A](throwable: Throwable): F[A]
   }
   object FxClient {
     def apply[F[_]: FxClient]: FxClient[F]         =
       implicitly[FxClient[F]]
     implicit def eftClientF[F[_]: Fx]: FxClient[F] = new FxClientF[F]
     final class FxClientF[F[_]: Fx] extends FxClient[F] {
-      override def eftOf[A](a: A): F[A] = effectOf(a)
-      override def of[A](a: A): F[A]    = pureOf(a)
-      override def unit: F[Unit]        = unitOf
+      override def eftOf[A](a: A): F[A]                 = effectOf[F](a)
+      override def of[A](a: A): F[A]                    = pureOf[F](a)
+      override def unit: F[Unit]                        = unitOf[F]
+      override def errOf[A](throwable: Throwable): F[A] = errorOf[F](throwable)
     }
   }
 
+  object TaskSpec {
+    import monix.execution.Scheduler.Implicits.global
+
+    @SuppressWarnings(Array("org.wartremover.warts.Any", "org.wartremover.warts.Nothing"))
+    def testAll: Property = for {
+      before <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("before")
+      after  <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).map(_ + before).log("after")
+    } yield {
+      @SuppressWarnings(Array("org.wartremover.warts.Var"))
+      var actual                  = before
+      @SuppressWarnings(Array("org.wartremover.warts.Var"))
+      var actual2                 = before
+      val testBefore              = actual ==== before
+      val testBefore2             = actual2 ==== before
+      val eftClient               = FxCtorClient[Task]
+      val effectConstructorClient = FxClient[Task]
+      val io                      =
+        for {
+          _  <- effectOf[Task]({ actual = after; () })
+          _  <- pureOf[Task]({ actual2 = after; () })
+          n  <- eftClient.eftOf(1)
+          n2 <- eftClient.of(n)
+          i  <- effectConstructorClient.eftOf(1)
+          i2 <- effectConstructorClient.of(1)
+          _  <- eftClient.unit
+          _  <- effectConstructorClient.unit
+        } yield ()
+      val testBeforeRun           = actual ==== before
+      val testBeforeRun2          = actual2 ==== before
+      io.runSyncUnsafe()
+      val testAfterRun            = actual ==== after
+      val testAfterRun2           = actual2 ==== after
+      Result.all(
+        List(
+          testBefore.log("testBefore"),
+          testBeforeRun.log("testBeforeRun"),
+          testAfterRun.log("testAfterRun"),
+          testBefore2.log("testBefore2"),
+          testBeforeRun2.log("testBeforeRun2"),
+          testAfterRun2.log("testAfterRun2")
+        )
+      )
+    }
+
+    @SuppressWarnings(Array("org.wartremover.warts.Any", "org.wartremover.warts.Nothing"))
+    def testEffectOf: Property = for {
+      before <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("before")
+      after  <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).map(_ + before).log("after")
+    } yield {
+      @SuppressWarnings(Array("org.wartremover.warts.Var"))
+      var actual        = before
+      val testBefore    = actual ==== before
+      val io            = effectOf[Task]({ actual = after; () })
+      val testBeforeRun = actual ==== before
+      io.runSyncUnsafe()
+      val testAfterRun  = actual ==== after
+      Result.all(
+        List(
+          testBefore.log("testBefore"),
+          testBeforeRun.log("testBeforeRun"),
+          testAfterRun.log("testAfterRun")
+        )
+      )
+    }
+
+    def testPureOf: Property = for {
+      before <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("before")
+      after  <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).map(_ + before).log("after")
+    } yield {
+      @SuppressWarnings(Array("org.wartremover.warts.Var"))
+      var actual        = before
+      val testBefore    = actual ==== before
+      val io            = pureOf[Task]({ actual = after; () })
+      val testBeforeRun = actual ==== after
+      io.runSyncUnsafe()
+      val testAfterRun  = actual ==== after
+      Result.all(
+        List(
+          testBefore.log("testBefore"),
+          testBeforeRun.log("testBeforeRun"),
+          testAfterRun.log("testAfterRun")
+        )
+      )
+    }
+
+    def testUnitOf: Result = {
+      val io             = unitOf[Task]
+      val expected: Unit = ()
+      val actual: Unit   = io.runSyncUnsafe()
+      actual ==== expected
+    }
+
+    @SuppressWarnings(Array("org.wartremover.warts.Any", "org.wartremover.warts.Nothing"))
+    def testErrorOf: Result = {
+      val expectedMessage = "This is a throwable test error."
+      val expectedError   = SomeThrowableError.message(expectedMessage)
+
+      val io = errorOf[Task][Unit](expectedError)
+      expectThrowable(io.runSyncUnsafe(), expectedError)
+    }
+
+  }
+
   object IoSpec {
-    import effectie.cats.Fx._
+    import cats.effect.IO
+    import effectie.monix.Fx._
 
     @SuppressWarnings(Array("org.wartremover.warts.Any", "org.wartremover.warts.Nothing"))
     def testAll: Property = for {
@@ -105,10 +215,10 @@ object EffectfulSpec extends Properties {
       Result.all(
         List(
           testBefore.log("testBefore"),
-          testBeforeRun.log("testBeforeRun"),
-          testAfterRun.log("testAfterRun"),
           testBefore2.log("testBefore2"),
+          testBeforeRun.log("testBeforeRun"),
           testBeforeRun2.log("testBeforeRun2"),
+          testAfterRun.log("testAfterRun"),
           testAfterRun2.log("testAfterRun2")
         )
       )
@@ -162,11 +272,12 @@ object EffectfulSpec extends Properties {
       actual ==== expected
     }
 
+    @SuppressWarnings(Array("org.wartremover.warts.Any", "org.wartremover.warts.Nothing"))
     def testErrorOf: Result = {
       val expectedMessage = "This is a throwable test error."
       val expectedError   = SomeThrowableError.message(expectedMessage)
 
-      val io = errorOf[IO](expectedError)
+      val io = errorOf[IO][Unit](expectedError)
       expectThrowable(io.unsafeRunSync(), expectedError)
     }
 
@@ -185,7 +296,8 @@ object EffectfulSpec extends Properties {
       after  <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).map(_ + before).log("after")
     } yield {
       implicit val executorService: ExecutorService = Executors.newFixedThreadPool(1)
-      implicit val ec: ExecutionContext             = ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
+      implicit val ec: ExecutionContext             =
+        ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
 
       @SuppressWarnings(Array("org.wartremover.warts.Var"))
       var actual                  = before
@@ -224,7 +336,8 @@ object EffectfulSpec extends Properties {
       after  <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).map(_ + before).log("after")
     } yield {
       implicit val executorService: ExecutorService = Executors.newFixedThreadPool(1)
-      implicit val ec: ExecutionContext             = ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
+      implicit val ec: ExecutionContext             =
+        ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
 
       @SuppressWarnings(Array("org.wartremover.warts.Var"))
       var actual               = before
@@ -245,7 +358,8 @@ object EffectfulSpec extends Properties {
       after  <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).map(_ + before).log("after")
     } yield {
       implicit val executorService: ExecutorService = Executors.newFixedThreadPool(1)
-      implicit val ec: ExecutionContext             = ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
+      implicit val ec: ExecutionContext             =
+        ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
 
       @SuppressWarnings(Array("org.wartremover.warts.Var"))
       var actual       = before
@@ -263,10 +377,11 @@ object EffectfulSpec extends Properties {
 
     def testUnitOf: Result = {
       implicit val executorService: ExecutorService = Executors.newFixedThreadPool(1)
-      implicit val ec: ExecutionContext             = ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
+      implicit val ec: ExecutionContext             =
+        ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
       val future                                    = unitOf[Future]
       val expected: Unit                            = ()
-      val actual: Unit                              = ConcurrentSupport.futureToValueAndTerminate(executorService, waitFor)(future)
+      val actual: Unit = ConcurrentSupport.futureToValueAndTerminate(executorService, waitFor)(future)
       actual ==== expected
     }
 
@@ -275,7 +390,8 @@ object EffectfulSpec extends Properties {
       val expectedError   = SomeThrowableError.message(expectedMessage)
 
       implicit val executorService: ExecutorService = Executors.newFixedThreadPool(1)
-      implicit val ec: ExecutionContext             = ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
+      implicit val ec: ExecutionContext             =
+        ConcurrentSupport.newExecutionContext(executorService, ErrorLogger.printlnExecutionContextErrorLogger)
 
       val future = errorOf[Future][Unit](expectedError)
       expectThrowable(ConcurrentSupport.futureToValueAndTerminate(executorService, waitFor)(future), expectedError)
@@ -284,7 +400,6 @@ object EffectfulSpec extends Properties {
   }
 
   object IdSpec {
-    import effectie.cats.Fx._
 
     def testAll: Property = for {
       before <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("before")
